Added task to update apt cache
Discovered that if you put this as your first role on a new box then the install will fail cause apt update hasn't occurred. If I put it after another role such as apache then it will work. It would seem that it might be best to put an apt update in all roles like this.